refactor: 기능 단위로 디렉토리 구조화
