+++ src/context/index.ts
@@ -1,3 +1,3 @@
-export * from "./themeContext";
-export * from "./userContext";
-export * from "./notificationContext";
+export * from "./notification";
+export * from "./user";
+export * from "./theme";
