+++ src/context/theme/useThemeContext.ts
@@ -1,5 +1,5 @@
 import { useContext } from "react";
-import { ThemeContext } from "../../context";
+import { ThemeContext } from "../theme";
 
 export const useThemeContext = () => {
   const context = useContext(ThemeContext);
