+++ src/context/user/useUserContext.ts
@@ -1,5 +1,5 @@
 import { useContext } from "react";
-import { UserContext } from "../../context";
+import { UserContext } from "./userContext";
 
 export const useUserContext = () => {
   const context = useContext(UserContext);
