+++ src/context/user/userContext.ts
@@ -1,5 +1,5 @@
 import { createContext } from "react";
-import { User } from "../types";
+import { User } from "../../types";
 
 interface UserContextType {
   user: User | null;
