+++ src/equalities/index.ts
