+++ src/equalities/shallowEquals.ts
