+++ src/utils/generateItems.ts
@@ -1,7 +1,3 @@
-export function renderLog(message: string) {
-  console.log(message);
-}
-
 // 대규모 데이터 생성 함수
 export const generateItems = (count: number, start = 0) => {
   const categories = ["전자기기", "의류", "도서", "식품"];
